requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
